added style for ComboBox, added ThumbnailRenderer and thumbnail generation to ResourceFile, added script support to IntPropertyEditor, added ComboBox to ResourcePropertyEditor, moved swap chain out of Renderer, added ResourceContext to Resource::save(..), fixed Graphics::CopyQueue's getter, added image capturing to RenderTarget, added Image saving

+++ Engine/src/graphics/mesh/SphereMesh.cpp
@@ -4,7 +4,7 @@
 
 namespace Ghurund {
     Status SphereMesh::init(Graphics& graphics, CommandList& commandList, unsigned int detail) {
-        float t = (1.0f + sqrt(5.0f)) / 2.0f;
+        float t = (1.0f + sqrtf(5.0f)) / 2.0f;
 
         Vertex triangleVertices[] = {
             {{-1.0, t, 0.0},{0,0}},
